* Makefile.am (EXTRA_DIST_M4): distribute m4/macros/libtool.m4.
* m4/macros/header_templates.m4: template for HAVE_EXTERN_MNTCTL.

* m4/macros/check_extern.m4: include <sys/vmount.h> on AIX
systems, so we can find extern definition for mntctl().

* conf/mtab/mtab_aix.c, conf/checkmount/checkmount_aix.c: onlu
define an xtern for mntctl() if needed.

* amd/ops_nfs.c (prime_nfs_fhandle_cache): never, ever, use C++
style comments in C code.  Some C compilers don't accept it.

* amd/info_ldap.c (amu_ldap_init, get_ldap_timestamp): force cast
of "time_t *ts" to u_long, to avoid complaints on AIX 5.2 with
xlC.

* amd/get_args.c: rename symbol conf_file to amu_conf_file to
avoid conflict with libldap.a from openldap-2.0.21 on AIX 5.2.
Oh, when will people who develop libraries learn to always prefix
their exported symbols with a library-specific name?

* configure.in: check for mntctl() extern on AIX systems.

* ltmain.sh: update from latst libtool-1.5.18, so we can properly
build on AIX 5.2/5.3 with cc and with gcc.

* m4/macros/libtool.m4: update from latest libtool-1.5.18.  This
time, and from now on, we must include libtool.m4 from the libtool
distribution itself.  That way we ensure that we use an ltmain.sh
that macthes the appropriate libtool M4 macros.  This makes
am-utils no longer dependent on having libtool installed on a
given system on which you run bootstrap.

+++ amd/info_ldap.c
@@ -37,7 +37,7 @@
  * SUCH DAMAGE.
  *
  *
- * $Id: info_ldap.c,v 1.26 2005/05/17 16:48:55 ezk Exp $
+ * $Id: info_ldap.c,v 1.27 2005/05/24 04:45:01 ezk Exp $
  *
  */
 
@@ -264,7 +264,7 @@ amu_ldap_init(mnt_map *m, char *map, time_t *ts)
   dlog("Bound to %s:%d\n", aldh->hostent->host, aldh->hostent->port);
   if (get_ldap_timestamp(aldh, map, ts))
     return (ENOENT);
-  dlog("Got timestamp for map %s: %ld\n", map, *ts);
+  dlog("Got timestamp for map %s: %ld\n", map, (u_long) *ts);
 
   return (0);
 }
@@ -418,7 +418,7 @@ get_ldap_timestamp(ALD *a, char *map, time_t *ts)
     }
     if (!*ts > 0) {
       plog(XLOG_USER, "Nonpositive timestamp %ld for map %s\n",
-	   *ts, map);
+	   (u_long) *ts, map);
       err = ENOENT;
     }
   } else {
@@ -429,7 +429,7 @@ get_ldap_timestamp(ALD *a, char *map, time_t *ts)
 
   ldap_value_free(vals);
   ldap_msgfree(res);
-  dlog("The timestamp for %s is %ld (err=%d)\n", map, *ts, err);
+  dlog("The timestamp for %s is %ld (err=%d)\n", map, (u_long) *ts, err);
   return (err);
 }
